Move encrypt/decrypt to one file
